Experimental GPS and GRID MAPPING Chat feature
If you want a simple chat version that works v097 is you best bet. This version requires testing especially in new GPS and Grid Map functionality. Currently this is most likely to work with same software to same software , it may not decode APRS RPR packets received at this time.